pyproject.toml
[project]
name = "mindstudio-agent"
version = "0.1.0"
description = "msAgent CLI aligned to the langrepl architecture"
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.11"

dependencies = [
    "aiofiles~=25.1.0",
    "aiosqlite~=0.22.1",
    "fastapi>=0.128.0",
    "greenlet~=3.3.0",
    "httpx>=0.28.0,<1.0.0",
    "json-repair>=0.54.3,<0.59.0",
    "jsonschema>=4.25.1,<4.27.0",
    "langchain~=1.2.0",
    "langchain-anthropic~=1.3.0",
    "langchain-aws>=1.1.1,<1.4.0",
    "langchain-community~=0.4.1",
    "langchain-core~=1.2.4",
    "langchain-deepseek~=1.0.1",
    "langchain-google-genai>=4.1.2,<4.3.0",
    "langchain-mcp-adapters~=0.2.1",
    "langchain-ollama~=1.0.1",
    "langchain-openai~=1.1.6",
    "langgraph~=1.0.5",
    "langgraph-checkpoint-sqlite~=3.0.1",
    "langgraph-cli[inmem]~=0.4.11",
    "langgraph-prebuilt~=1.0.5",
    "markdown-it-py~=4.0.0",
    "mdformat~=1.0.0",
    "mcp>=1.13.0,<2.0.0",
    "packaging>=24.0,<26.0",
    "pathspec>=0.12.1,<1.1.0",
    "prompt-toolkit~=3.0.52",
    "pydantic~=2.12.5",
    "pydantic-settings>=2.12,<2.14",
    "pygments-cache~=0.1.3",
    "python-dotenv~=1.2.1",
    "python-multipart~=0.0.21",
    "pyyaml~=6.0.3",
    "rich>=14.2,<14.4",
    "tiktoken>=0.9.0,<1.0.0",
    "trafilatura~=2.0.0",
    "pyseccomp~=0.1.2; sys_platform == 'linux'",
]

[project.scripts]
msagent = "msagent.cli.bootstrap.app:cli"

[tool.mypy]
python_version = "3.11"
mypy_path = ["src"]

[[tool.mypy.overrides]]
module = [
    "langchain_mcp_adapters.*",
    "pygments_cache.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "msagent.mcp.factory"
disable_error_code = ["assignment", "index", "typeddict-unknown-key", "arg-type"]

[tool.pytest.ini_options]
pythonpath = [".", "src"]
testpaths = ["tests"]
filterwarnings = [
    "ignore:LangSmith now uses UUID v7:UserWarning:pydantic.v1.main",
    "ignore:Using fallback GPT-2 tokenizer.*:UserWarning",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/msagent", "resources"]
force-include = { "src/msagent/py.typed" = "msagent/py.typed", "skills" = "msagent/skills" }

[dependency-groups]
dev = [
    "black>=25.12,<26.2",
    "mypy~=1.19.1",
    "pytest~=9.0.2",
    "pytest-asyncio~=1.3.0",
    "ruff>=0.12.0,<0.14.0",
]
